# Reduction performance tweaks