--- a/src/reduction/bit_depth.rs
+++ b/src/reduction/bit_depth.rs
@@ -16,13 +16,13 @@ pub fn reduced_bit_depth_16_to_8(png: &PngImage, force_scale: bool) -> Option<Pn
     }
 
     // Reduce from 16 to 8 bits per channel per pixel
-    if png.data.chunks(2).any(|pair| pair[0] != pair[1]) {
+    if png.data.chunks_exact(2).any(|pair| pair[0] != pair[1]) {
         // Can't reduce
         return None;
     }
 
     Some(PngImage {
-        data: png.data.iter().step_by(2).copied().collect(),
+        data: png.data.chunks_exact(2).map(|pair| pair[0]).collect(),
         ihdr: IhdrData {
             color_type: png.ihdr.color_type.clone(),
             bit_depth: BitDepth::Eight,
@@ -41,7 +41,7 @@ pub fn scaled_bit_depth_16_to_8(png: &PngImage) -> Option<PngImage> {
     // Reduce from 16 to 8 bits per channel per pixel by scaling when necessary
     let data = png
         .data
-        .chunks(2)
+        .chunks_exact(2)
         .map(|pair| {
             if pair[0] == pair[1] {
                 return pair[0];