Two key changes here:
- Removed some defunct reduction benches and added some more alpha reduction benches.
- Changed use of `chunks()` to `chunks_exact()` where appropriate to improve performance. (As a byproduct, this also fixes a potential crash if a malformed file has a PLTE chunk that isn't a multiple of 3.)