@andrews05 2025-02-03
The `chunks_exact` makes a big difference to the 16-to-8 benchmarks. (Still not appreciable in actual use though.)
Before:
```
test reductions_16_to_8_bits                        ... bench:     352,770.85 ns/iter (+/- 11,464.83)
test reductions_16_to_8_bits_scaled                 ... bench:     748,591.70 ns/iter (+/- 7,528.53)
```
After
```
test reductions_16_to_8_bits                        ... bench:     200,931.25 ns/iter (+/- 4,184.77)
test reductions_16_to_8_bits_scaled                 ... bench:     116,423.44 ns/iter (+/- 652.00)
```